follow payment element pattern
