+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Checkout/Express Checkout Element/ExpressCheckoutElementView.swift
@@ -5,6 +5,7 @@
 //  Created by Joyce Qin on 7/22/26.
 //
 
+import Combine
 import SwiftUI
 
 /// A SwiftUI view that displays wallet payment buttons (Apple Pay, Link).
@@ -23,6 +24,28 @@ public struct ExpressCheckoutElementView: View {
             ExpressCheckoutElementUIViewRepresentable(uiView: viewModel.uiView)
                 .fixedSize(horizontal: false, vertical: true)
         }
+    }
+}
+
+/// Bridges ExpressCheckoutElement's UIKit state into SwiftUI without retaining Checkout.
+@MainActor
+final class ExpressCheckoutElementViewModel: ObservableObject {
+    let uiView: ExpressCheckoutElementUIView
+    @Published var isAvailable: Bool
+
+    private var sessionCancellable: AnyCancellable?
+
+    init(checkout: Checkout, uiView: ExpressCheckoutElementUIView) {
+        self.uiView = uiView
+        // TODO: Derive from session (e.g. session.isExpressCheckoutElementAvailable)
+        self.isAvailable = true
+        sessionCancellable = checkout.$session
+            .dropFirst()
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] _ in
+                // TODO: Derive from session (e.g. session.isExpressCheckoutElementAvailable)
+                self?.isAvailable = true
+            }
     }
 }
